server/requirements.txt
# Speech Recognition & Transcription
openai-whisper==20231117

# Speaker Diarization
pyannote.audio==3.1.1

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Audio Processing
soundfile==0.12.1
librosa==0.10.1

# Utilities
python-dotenv==1.0.0
